bug: getAdapter when called from a Resource fails
When the getAdapter method on a Resource is called, the first argument is always the resource name as set here: https://github.com/js-data/js-data/blob/master/src/datastore/sync_methods/defineResource.js#L241-L248.

The fix I suggest is simply to check for the argument length, and set options accordingly. It's a little ugly but it works. What do you think? Is there a better way to achieve this?