DS#find() returns items cached with DS#inject()
DS#find() currently only returns the cached item early only if a
previous request was made to fetch the item. If the item was added using
DS#inject(), DS#find() will make a request anyways.

This adds a findStrictCache option that can be set to false to make
DS#find() return the item directly instead.